aibricks/aura_dataprep/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# aura_dataprep/scripts/aura_ecg_detector.py: 5
biosppy == 0.7.3

# aura_dataprep/aura_dataprep_pb2.py: 5,6,7,8
google_auth == 2.1.0

# aura_dataprep/aura_dataprep.py: 1
# aura_dataprep/aura_dataprep_pb2_grpc.py: 3
# aura_dataprep/test_aura_dataprep.py: 2
grpcio == 1.41.1

# aura_dataprep/scripts/aura_features_computation.py: 6
hrv_analysis == 1.0.4

# aura_dataprep/scripts/aura_ecg_detector.py: 1
# aura_dataprep/scripts/aura_features_computation.py: 3
numpy == 1.19.5

# aura_dataprep/scripts/aura_ecg_detector.py: 8
pyEDFlib == 0.1.22

# aura_dataprep/scripts/aura_ecg_detector.py: 3
py_ecg_detectors == 1.0.2

# aura_dataprep/test_aura_dataprep.py: 6
pytest == 6.2.4

# aura_dataprep/scripts/aura_features_computation.py: 9
scipy == 1.7.1

# aura_dataprep/scripts/aura_ecg_detector.py: 4
wfdb == 3.4.0
